backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
pytest==7.4.4
pytest-json-report==1.5.0
Flask-SQLAlchemy
flask-cors
google-generativeai
python-dotenv
